requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.52.0
django-allauth-2fa==0.10.0
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-environ==0.9.0
django-otp==1.1.4
django-phonenumber-field==7.0.2
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.3.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
